+++ extensions/blog/src/BlogExtension.php
@@ -6,7 +6,7 @@ use Pagekit\Application as App;
 use Pagekit\Blog\Content\ReadmorePlugin;
 use Pagekit\Blog\Event\CommentListener;
 use Pagekit\Blog\Event\RouteListener;
-use Pagekit\Extension\Extension;
+use Pagekit\System\Extension;
 use Pagekit\System\Event\LinkEvent;
 use Pagekit\System\Event\LocaleEvent;
 use Pagekit\System\Event\TmplEvent;
