+++ extensions/page/src/PageExtension.php
@@ -3,7 +3,7 @@
 namespace Pagekit\Page;
 
 use Pagekit\Application as App;
-use Pagekit\Extension\Extension;
+use Pagekit\System\Extension;
 use Pagekit\Page\Event\NodeListener;
 use Pagekit\System\Event\LinkEvent;
 use Pagekit\System\Event\LocaleEvent;
